requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
requests==2.31.0
pydantic==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==3.2.2